+++ apps/senstate-dashboard/src/app/components/tab-content/watchers-list/watchers-list.component.ts
@@ -1,7 +1,7 @@
-import {Component, OnInit, ChangeDetectionStrategy, Input} from '@angular/core';
+import {Component, OnInit, ChangeDetectionStrategy, Input, TrackByFunction} from '@angular/core';
 import {Observable} from "rxjs";
 import {WatcherMeta} from "@senstate/client";
-import {HubService} from "../../../state/hub.service";
+import {GroupedWatchers, HubService} from "../../../state/hub.service";
 
 @Component({
   selector: 'senstate-watchers-list',
@@ -14,16 +14,19 @@ export class WatchersListComponent implements OnInit {
   @Input()
   public appId: string;
 
-  public watchers$: Observable<WatcherMeta[]>;
+  public watchers$: Observable<GroupedWatchers[]>;
+
+  public trackByWatcherFunc: TrackByFunction<WatcherMeta> = (index, item) => {
+    return item.watchId;
+  };
+
+  public trackByGroupFunc: TrackByFunction<GroupedWatchers> = (index, item) => {
+    return item.key;
+  };
 
   constructor(private hub: HubService) { }
 
   ngOnInit() {
-    this.watchers$ = this.hub.getWatchersByApp$(this.appId)
-  }
-
-
-  public trackByWatcherFunc (tagObj: any) {
-    return tagObj.watchId;
+    this.watchers$ = this.hub.getGroupedWatchersByApp$(this.appId);
   }
 }